requirements.txt
# Python dependencies
# Install:  pip install -r requirements.txt

# HTTP + env
requests
python-dotenv

# Data manipulation (used by transform/build_stg_participants_csv.py)
pandas

# Warehouse engine (file-based analytics DB)
duckdb

# Transformation layer
dbt-core
dbt-duckdb
